STAR-406: End-to-end Row Awareness for SAI
STAR-406 Remove TotalPartitionReads metrics

Counting visited partitions was incorrect, because the counter
was advanced on each row, not each distinct (not-previously-seen)
partition.

The index searcher returns rows, not partitions, and generally there
is no guarantee that the order of the returned rows is the same as the
order of partitions. When we add ORDER BY a different column, the order of
iterating rows will definitely not match the order of partitions.
This means the same partition can be visited more than once.
Hence counting the number of distinct partitions visited would be costly,
so we decided to leave it out.

STAR-406 Port verifyCleanupFailedTokenOffsetFiles test to row aware

The test no longer makes sense, because there are no token and
offset components any more. They were replaced by
the primary key map (PRIMARY_KEYS component).
This commit replaces the test with a similar
test for cleanup after a failed write of PRIMARY_KEYS.

STAR-406 Fix PostingsTest

A part of STAR-348 fix (moved postings decoded to mark on postings
reader close (#87)) was probably lost during STAR-406 rebase.

Fixes:

java.lang.AssertionError:
Expected :6
Actual   :0
<Click to see difference>
	at __randomizedtesting.SeedInfo.seed([BA4168263DC903C4:775368E19EE347AF]:0)
	at org.junit.Assert.fail(Assert.java:89)
	at org.junit.Assert.failNotEquals(Assert.java:835)
	at org.junit.Assert.assertEquals(Assert.java:647)
	at org.junit.Assert.assertEquals(Assert.java:633)
	at org.apache.cassandra.index.sai.disk.v1.PostingsTest.testSingleBlockPostingList(PostingsTest.java:78)

STAR-848 Use a flat list of offsets to implement rowId to primary key… (#233)

* STAR-848 Use a flat list of offsets to implement rowId to primary key map

Instead of storing primary keys in a
KD-tree, we write them into a flat file one after another and
just record offsets to them. We can do this because we know the primary keys
come in the correct order.

Advantages:
* code is much simpler
* by not sorting the keys in the KD-tree, we'll never mess up their order
* looking up a primary key should take approximately constant time

Downsides:
* we lost prefix compression of the keys; to be fixed later

Limitations of the implementation:
* currently we do 2 lookups of offsets per one rowId lookup in order
  to get the length of the primary key before deserializing;
  I believe this can be done better
* there is additional temp byte array allocation on the heap - I also
  believe this could be removed if we rethink the key deserialization logic

* STAR-848 Add header and footer to PrimaryKeys component

* STAR-848 Fix SSTableComponentsTest

We no longer use memtable tries in SSTableComponentWriter
so the test needed to be adapted a bit.

* STAR-848 Add PRIMARY_KEY_OFFSETS to PER_SSTABLE_COMPONENTS

* STAR-848 Fix number of expected components in tests

* STAR-848 Reenable column component validation

STAR-867 Fix resource management errors related to PrimaryKeyMap (#244)

Fixed FileHandle leaks caused by lack of proper closing of PrimaryKeyMap
in various places. Removed sharing of primaryKeyOffsets
between copies of PrimaryKeyMap.