STAR-561 selectors should contain elements with same type (#169)
* selectors should contain elements with same type

Expressions like "SELECT [(CAST(pk AS BIGINT), CAST(ck AS BIGINT)), (t, WRITETIME(t))] FROM %s"
can't be deserialized. If we try to extract the result by calling
getRows we get:
org.apache.cassandra.serializers.MarshalException: Expected 8 or 0 byte long

This commit makes sure that all collection elements have the
same type.

Co-authored-by: Zhao Yang <zhao.yang@datastax.com>

* test for Maps.getExactMapTypeIfKnow

Maps are a bit different than Lists and Sets. Maps can't be selected
without a type hint, which makes it impossible to exercise query execution
path with `getExactMapTypeIfKnow`. That is why tests for "map SELECTs" are
missing TermSelectionTest.java.
The previous fixed `Maps.getExactMapTypeIfKnow` but it can't be tested in
a similar way as for Lists and Sets. That's why commit adds a separate
unit tests for `Maps.getExactMapTypeIfKnow`.

Co-authored-by: Zhao Yang <zhao.yang@datastax.com>