STAR-580 replace polymorphic calls with a switch
The original description:

In profiling JMH workload ReadWide a bottleneck has been identified
in the compareTo call made for AbstractType subclasses. The
comparison code seem overtly generic. For the numerical types in
particular, it is better to specialize.

This ported commit replaces polymorphic 'compare' calls with a
switch. A single entry point for comparison has been created -
AbstractType.compare.
Primitive types were extracted to a separate subgroup with
static calls.