STAR-766 fix testGetScannerForNoIntersectingRanges
Since CASSANDRA-5249 getScanner returns
EmptySSTableScanner when there are no intersecting ranges.
Empty scanner always returns false from hasNext().

Why testGetScannerForNoIntersectingRanges worked although
hasNext returns false? Other test created a second
sstable for which the scanner was not empty and thus
the assertion passed. In other words, the test failed when
executed in isolation.

This commit creates a dedicated table for the test to avoid
other test interference. Additionally the number of sstables
is verified in the test.
Instead of an assertion based on hasNext, an assertion on
EmptySSTableScanner is made.