STAR-748: Fix Scrubber so that it can work in case of broken index (#193)
Fix scrubber so that it can work with broken indexes.

If the index is broken, but the data file is ok, we can continue scrubbing and the index will be rebuilt. With this patch it will work even if we cannot access the initial position in the index. Also, the corrupted file will be obsoleted before finishing the rewriter so that it will not be attempted to move starts (and thus fail at the end as move starts requires correct indexes).

There are also some explanations added to how canonical set of sstables is constructed and why

Co-authored-by: Branimir Lambov <branimir.lambov@datastax.com>