STAR-801: Refactor the logic for loading and recreating bf
The logic for deciding when to load and when to recreate a bf was separately defined for bti and big. Also, for big the logic has changed in STAR-745 and some tests were added to verify it.

In this commit the logic has been extracted to top SSTableReader level and used consistently for both bti and big.

Post STAR-247+STAR-745 change

It should fix the test failure in SSTableReaderTest