STAR-801: Adjust the failing assertion in SSTablesIteratedTest.testNonCompactTableWithStaticColumnValueMissingAndMulticellColumn
The test was introduced recently in OSS to verify the correctness when dealing with static columns.

The failing assertion expects 3 sstables to be read when selecting non-static value, which in fact is present only in one sstable. It is correct in OSS that we read 3 sstables as all of them touch that partition however they do not cover that particular clustering key which we query for. The difference results from read path optimizations we applied in STAR-1. That optimization improves filtering of sstables by making it more precise so that more sstables can be skipped. The optimization comes into play in this test case.

(cherry picked from commit 1c24e77b52ade9e4bb94016649b9492b7a918cdc)