STAR-821 Add a way to flatten WHERE expression tree
Collapsing levels with the same operator makes reasoning
about expression properties much simpler. AND and OR are
commutative and associative, so nesting is just a distraction.

Additionally refactored the expression tree container nodes
to be truly immutable and fixed a subtle error in rename() that
caused an OR node to be replaced by an AND node.

(cherry picked from commit f2ace24ba6f797049752e1e750ec2f548327ee0f)