STAR-406 Remove TotalPartitionReads metrics
Counting visited partitions was incorrect, because the counter
was advanced on each row, not each distinct (not-previously-seen)
partition.

The index searcher returns rows, not partitions, and generally there
is no guarantee that the order of the returned rows is the same as the
order of partitions. When we add ORDER BY a different column, the order of
iterating rows will definitely not match the order of partitions.
This means the same partition can be visited more than once.
Hence counting the number of distinct partitions visited would be costly,
so we decided to leave it out.