STAR-823: Fix deadlock in CompactionManager (#230)
STAR-823: Refactor background compactions

CompactionManager.BackgroundCompactionCandidate task is scheduled on
compaction executor and when it detects there are compaction tasks
to run, it starts each compaction task as a separate job on the same
compaction executor and blocks until all tasks are finished.

When the pool size for the executor is n, and there are n background
tasks submitted in parallel, and all of them find that there are some
compactions to run, they will schedule them and block until the tasks
are finished. Though, the tasks cannot start because the pool is full
- we have n background tasks there waiting for the compaction tasks
that cannot start.

Another thing (perhaps minor) is that we use getActiveCount() on 
the executor to check how many tasks it is currently running, 
and based on that information either schedule new tasks or not. 
The problem with this method is that it returns an approximate 
result and should not be used for making such decisions.

To address those problems, running of background compactions was
refactored. The whole logic for background compactions was extracted
into a distinct class BackgroundCompactionsRunner. It allows flagging
CFSs for compaction and schedules scans through the flagged CFSs
on a dedicated executor so that scans and compaction tasks are no
longer sharing the same executor.

Co-authored-by: Branimir Lambov <branimir.lambov@datastax.com>