STAR-844 Read/Write hooks (#250)
This is a port of
- https://github.com/riptano/bdp/commit/b6f0a18cb832c62f05cdcbd9cdcc2923f2fa727f
- https://github.com/riptano/bdp/pull/19468

The first change set introduces QueryInfoTracker (QIT) interface and
hooks it to StorageProxy. The second adds ClientState to the interface.

The original QIT utilizes ReadReconciliationObserver in the ReadTracker
paths. Only onRow, onPartition and queried callbacks are utilized by
CNDB and thus only these methods are ported to Converged Cassandra
(CC). The callbacks are a bit different tho:
- The callback methods are added directly to ReadTracker as CC doesn't
have ReadReconciliationObserver. The class was added as a part of
NodeSync effort and it is rather superfluous. Porting the whole class would
add unnecessary complexity. Adding the required methods directly to the
ReadTracker makes the interface cleaner and easier to understand.
- CC operates on ReplicaPlans instead of plain host lists, that is why queried
was changed to onReplicaPlan.

(cherry picked from commit c32e91ff0fcb06bc4b7637670627c65dfa27e21c)