STAR-790: Fix changes application order in SchemaKeyspace
Sometimes it happens we replace the keyspace. In this case the keyspace with the same name is in dropped and created sets. If we first try to create it, we would just replace the existing one and the remove it, which is not what we want.

(cherry picked from commit 6cf56ee310f4325fad229faa04bf0c101388819c)