STAR-790: Add SchemaUpdateHandlerFactoryProvider
Factory is used to create the right instance depending on the context - whether we are online or not (server mode / tool or client mode). Provider is for creating the factory so that we can switch the implementation by providing a system property. This can be also useful for testing.

(cherry picked from commit 7c8227dd4560073ad5a32d4871bee1edf5f0fd1e)