CASSANDRA-17044: Extracted LocalKeyspaces class
LocalKeyspaces holds metadata of the system local keyspaces, namely system and system_schema.
Metadata of those keyspaces are immutable, they are not synced with other nodes and not stored. As such, we want to separate them from metadata of the keyspaces which are modifiable and synced between nodes, hence called sharedKeyspaces.

There a couple of minor refactorings following that change - some query methods in SchemaManager were renamed and refactored to return immutable Keyspaces object rather than ordinary collections of KeyspaceMetadata or names.