STAR-790: Fix duplicated client warnings in schema statements
This seemed to be screwed a bit. In just two schema alteration statements we collected client warning which were captured during the transformation into a local collection. I guess it was done because the transformation was being done in a different stage (migration) and client warnings collected in that stage would not be present in the stage where the query was executed. Then, the client warnings are retrieved using clientWarnings method and added to the captured client warnings in the stage which is executing the query. Once we are not executing explicit transformations in migration stage (the methods are synchronized anyways and this was a blocking operation) the client warnings got duplicated - we collected them and then added again. This weird mechanism was implemented only in two schema alteration statements - I wonder if for other ones, the client warnings could simply get lost before this change.

(cherry picked from commit 078de83a504e5915fe65d4ab6ca284313aa842c3)