STAR-1207: Open iterators lazily in StorageAttachedIndexSearcher (#385)
STAR-1207: Open iterators lazily in StorageAttachedIndexSearcher

In STAR-1180, the StorageAttachedIndexSearcher was modified
so that each computeNext() returns an iterator for a different
partition. This fixed the correctness issue, but the implementation
was far suboptimal - it required fetching all data of a partition
eagerly before the final merged partition iterator could be returned.
In case of extremely wide partitions that behavior could destabilize
the server.

This commit gets rid of eager creation of the row iterators.
It keeps only one iterator open at a time, and opens new iterators only
when the previous one runs out of data. The data retrieval commands
are issued lazily, while the returned iterator gets iterated.

In addition, the code has been tidied up, organized and documented
in order to improve readability and future maintenance.