STAR-1180: Fix premature skip to the next keyRange
If queryController.select(key) returned false,
it was possible that the loop in nextKeyInRange advanced
the currentKeyRange prematurely, potentially skipping a bunch
of keys which led to missing some rows in the result of a query.

The code has been simplified by taking the
check for queryController to a separate method.

(cherry picked from commit f13d8142f7fd217d99d32cf7744695af2d60b616)