STAR-540: Fix issues with SAI queryability and status propagation (#410)
STAR-540: Fix multiple issues with SAI queryability and SAI status propagation

* STAR-540: Don't query index in unknown state

Allow index to be queried only if its status is known
to be BUILD_SUCCEEDED. For all the other states,
including the UNKNOWN state, it is inherently dangerous
to query such index. A query to an index that's
in the middle of building process may return incomplete
results.

Fixes flakyness of QueryWriteLifecycleTest manifested as:

    junit.framework.AssertionFailedError: expected:<[MS, AL, TN, UT, KY, LA]> but was:<[MS, AL, TN, KY, LA]>
	    at org.apache.cassandra.index.sai.cql.IndexQuerySupport$BaseQuerySet.validate(IndexQuerySupport.java:568)
	    at org.apache.cassandra.index.sai.cql.IndexQuerySupport$BaseQuerySet.rangeQuery(IndexQuerySupport.java:551)
	    at org.apache.cassandra.index.sai.cql.IndexQuerySupport$BaseQuerySet.execute(IndexQuerySupport.java:354)
	    at org.apache.cassandra.index.sai.cql.IndexQuerySupport.executeQueries(IndexQuerySupport.java:215)
	    at org.apache.cassandra.index.sai.cql.IndexQuerySupport.writeLifecycle(IndexQuerySupport.java:97)
	    at org.apache.cassandra.distributed.test.sai.QueryWriteLifecycleTest.testWriteLifecycle(QueryWriteLifecycleTest.java:30)

Additionally, this commit fixes the code for waiting for index readiness in SAIUtil.
The original code ignored the fact that if the schema with the new index
hasn't been propagated to the other node yet, that node does not know anything
about the index and doesn't return its status. Unfortunately a
lack of status was considered as "index queryable" and the waiting
didn't happen.

SAIUtil.waitForIndexQueryable has been modified to rely on index status from
gossip. This way it can see the exact same state that a coordinator
uses to determine index availability. The downside is that now some tests
need to run with gossip enabled.

* STAR-540: Fix SAI index status propagation issues

This commit fixes sevaral issues related to how index
status is announced in the cluster through Gossip.

1. When a new node gets bootstrapped, the status of indexes
is now properly announced through gossip in all cases,
even if indexes didn't need building because they had been
already streamed.

2. Index status is now received from a peer even
before the peer goes to NORMAL state, while the peer is
bootstrapping. This way, when the new node transistions to
the NORMAL state, we already have up-to-date information
on its indexes and the newly joined node can become
immediately queryable.

3. Additional DEBUG logging has been added to display
index statuses received from peers.

4. TestBaseImpl#bootstrapAndJoinNode now additionally waits
for the other nodes to notice the newly joined node switch to
NORMAL state. This prevents some problems caused by tokens
suddenly moving later while the test code is running.
Fixes flakyness in NativeIndexDDLTest. That test sometimes
didn't do keyspace cleanup because the first node had
outdated token ring information and incorrectly
believed to be an owner of the whole ring.

(cherry picked from commit c164af08887be8e180499a901c65c95dc71ed94d)