STAR-1235: Fix deadlock (STAR-996) (#408)
The deadlock happens when we want to flush a change as part of Nodes update, when we update some Nodes value which requires immediate flushing. Flushing triggers compaction, and compaction tries to update some endpoint state in Gossiper (sstable versions). Endpoint state in Gossiper is automatically synced with Nodes, so any update on endpoint state makes a synchronized request to update the relevant Nodes value. This way we attempt to update Nodes inside updating Nodes which causes a deadlock.

In general we should avoid such situations, but in this particular case, sstable versions is not maintained by Nodes and therefore we do not have sync such change with Nodes (in fact, it would be empty update). The fix consists of determining if updates on Nodes are really needed before actually calling synchronous update on Nodes.

I've made updates in endpoint state synchronous because otherwise we could have a situation when gossiper reports different values than those represented by Nodes (or by system tables before Nodes were introduced).