STAR-1242 STAR-1243 Extend compaction pick for CNBD need (#413)
This patch allows to set the transaction UUID directly when a
compaction pick is created, which means that in CNDB we can use it as a
key for compaction tasks, making debugging of compactions much easier
(since we can grep the same UUID). Currently, we're using the name of
one of the sstables as the unique key of compaction tasks, but this is
problematic when analyzing the logs, especially after a restart.

Also, the transaction UUID was already a property of the compaction
pick, but it was nullable and volatile. It is now final, which is an
improvement, and a boolean now signals whether the pick has been
submitted or not.

Port of CNDB-2291 and of minor part from CNDB-3376

Co-authored-by: Stefania Alborghetti
(cherry picked from commit d0265d09e0df0815e764e3fa01d2ebb056439349)
(cherry picked from commit e0e6964f52c052a0872b7fccf21e546e5f516e40)
(cherry picked from commit f7e356dd3b2011a1f62da3e7d247788e8cef7dea)