STAR-1261: Don't try to establish TLSv1.1 connections in test
TLSv1.1 has been disabled globally in OpenJDK and various
other JVMs in April 2021. Although it is possible to unlock it
by editing the java.security file, the test cannot unlock
it by itself. Settings in the java.security file are stronger
than the settings passed at runtime when building the SslContext.

Fortunately, support for TLSv1.3 landed in OpenJDK 8 some time
ago (but is not selected by default), so we can use that protocol
to test the protocol options.

Fixes InternodeEncryptionOptionsTest.negotiatedProtocolMustBeAcceptedProtocolTest.