STAR-1694 Expose buckets in UCS and accept sstables for direct classification
This commit changes the API of UCS as follows:

- The Bucket inner class is now public
- The method for extracting shards with buckets is now public, and it
  accepts a custom list of sstables

These changes are required for CNDB, so that we can classify all live sstables
and visualize their corresponding shards and buckets in a diagnostic tool such as
Autobot.

The comments for warnIfSizeAbove have been clarified and moved to the method Javadoc.

Port of CNDB-4385