STAR-1598 port CNDB-3932 (#509)
If a user creates a prepared statement for a table and then enables
the CDC param, the prepared statement must be invalidated.

If we don't invalidate the prepared statement it'll be executed using
the old TableMetadata object and take the non-CDC write path and go
directly to CNDB writers instead of via Pulsar topics.

(cherry picked from commit 1bed2e7ff4588a5774fd0a46794209fd14997590)