STAR-1502: Fix a bug in scrubber
There was an uncovered case when there is just one item in the index and the data file is completely broken. The scrubber falls into infinite loop trying to move to the next partition using index. But since the index is exhausted, it cannot make any progress and also it was not marking that moving forward makes no sense.

(cherry picked from commit cf81e17e067035362bd978d56c1db0919dd46d11)