STAR-1443: Fix VerifyTest (#533)
Basically there were two tests which verify the same thing, but one was intended to use BTI and the other one BIG format.
Since by default we run with BTI, the BIG test explicitly sets the property to use BIG format. However, we run the full tests suite for both BTI and BIG, so when we run with BIG by default, the BTI fails (because it has no explicit configuration and assumes we run with BTI as default). The fix is simple - just have a single test, which just uses the default sstable format, it will use BTI in one test run and BIG in the other. Those squashed tests didn't differ by anything but assertion on the created sstable format.