STAR-1691: Port more bits of CNDB-4090 (#548)
In STAR-1335 we ported most of CNDB-4090 but we missed out a call to
StorageProvider.invalidateFileSystemCache() which is required to
invalidate the remote storage cache in CNDB whenever we encounter
corruption.

This was discovered because the RemoteFileCacheCorruptedPageTest unit
test is failing.

(cherry picked from commit 116f5d55d5a35ecb66613712da29d2840dd8f974)