STAR-1746: Don't throw an exception if we fail to unload a CF (#568)
This is a partial port of CNDB-3190 and is required to make the CNDB unit test,
TableDataDeletionTest.testDropTableAllDMsUploadFailed, work.

The main idea behind CNDB-3190 is that we don't want to allow any
exceptions that occur when unloading a CF to interfere with tenant
unassignment or make the schema inconsistent.

(cherry picked from commit c4845fcfa151634721f359de6522b509622de39c)