STAR-550 Handle SAI AbortedOperationException
AbortedOperationException is thrown by SAI when index search
hits a timeout. Now instead of allowing this exception to
bubble up to the top and be eventually logged as error, we catch
and swallow it in the InboundSink after creating the query response.
Additionally now we also we set a proper error code (TIMEOUT)
in the response, so the client has a hint on what happened.