port: DSP-23333: Limit row size during deserialization (#20725)
Do not allow reading more bytes than the declared row size in the row preamble when reading from the sstable. This is achieved by wrapping the input source with TrackedDataInputPlus which has now a new feature of limiting the read data.

When the sstable is corrupted, we can try to read much more data for a row than we should. In some cases, it can lead to OOMs when we read malformed cell counts and sizes.