More sophisticated approach to estimatedNodesVisited:
(1) Incorporate the count of permitted nodes, which has a major
effect when small (because we have to search a large part of the
graph before finding enough candidates).

(2) All tested datasets are ~linear wrt the new prediction, but
with different constants, so track actual vs predicted ratio and
use that to improve the prediction.